Updating the dataBase class
Updating the dataBase class by making the urlsObj always the same as urls.json file.
also: adding a test in the MakeNewShortenedUrl(url) method that check if theres already a shortened url to the given url.

+++ DBClass.js
@@ -3,25 +3,40 @@ const fs = require("fs");
 // Class for DataBase
 class DataBase {
   constructor() {
-    this.urlsArr = [];
+    fs.readFile("./urls.json", (err, data) => {
+      if (err) {
+        throw new Error(`message: ${err}`);
+      } else {
+        this.urlsObj = JSON.parse(data);
+      }
+    });
   }
 
   MakeNewShortenedUrl(url) {
+    const isExistUrl = this.urlsObj.urlsArr.filter(
+      (obj) => obj.originalUrl === url
+    );
+    if (isExistUrl.length > 0) {
+      return isExistUrl;
+    }
     const newUrlObject = {};
     newUrlObject.creationDate = dateToSqlFormat();
     newUrlObject.redirectCount = 0;
     newUrlObject.originalUrl = url;
     newUrlObject["shorturl-id"] = shortUrlGenerator();
-    this.urlsArr.push(newUrlObject);
+    this.urlsObj.urlsArr.push(newUrlObject);
     fs.writeFile(
       "./urls.json",
-      JSON.stringify(this.urlsArr, null, 4),
+      JSON.stringify(this.urlsObj, null, 4),
       (err) => {
         if (err) throw new Error(`message: ${err}`);
       }
     );
+    return newUrlObject;
   }
 }
+
+const dataBase = new DataBase();
 
 function dateToSqlFormat() {
   let timeCreation = new Date();
@@ -76,3 +91,5 @@ function shortUrlGenerator() {
   }
   return string;
 }
+
+module.exports = dataBase;
